# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SIKURI SI - Sistem Informasi Kurikulum Program Studi S1 Sistem Informasi
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mata-kuliah:
    get:
      operationId: listMataKuliah
      tags: [mata-kuliah]
      summary: Daftar semua mata kuliah
      parameters:
        - name: semester
          in: query
          schema:
            type: ["integer", "null"]
        - name: tipe
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MataKuliah"
    post:
      operationId: createMataKuliah
      tags: [mata-kuliah]
      summary: Tambah mata kuliah baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MataKuliahInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MataKuliah"
  /mata-kuliah/{id}:
    get:
      operationId: getMataKuliah
      tags: [mata-kuliah]
      summary: Detail satu mata kuliah
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MataKuliah"
        "404":
          description: Not found
    patch:
      operationId: updateMataKuliah
      tags: [mata-kuliah]
      summary: Update mata kuliah
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MataKuliahUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MataKuliah"
    delete:
      operationId: deleteMataKuliah
      tags: [mata-kuliah]
      summary: Hapus mata kuliah
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CPL PRODI ───────────────────────────────────────────────────────────────
  /cpl:
    get:
      operationId: listCpl
      tags: [cpl]
      summary: Daftar CPL Prodi
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CplProdi"
    post:
      operationId: createCpl
      tags: [cpl]
      summary: Tambah CPL baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CplProdiInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CplProdi"
  /cpl/{id}:
    get:
      operationId: getCpl
      tags: [cpl]
      summary: Detail CPL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CplProdi"
    patch:
      operationId: updateCpl
      tags: [cpl]
      summary: Update CPL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CplProdiUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CplProdi"
    delete:
      operationId: deleteCpl
      tags: [cpl]
      summary: Hapus CPL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CPMK ────────────────────────────────────────────────────────────────────
  /cpmk:
    get:
      operationId: listCpmk
      tags: [cpmk]
      summary: Daftar semua CPMK
      parameters:
        - name: cpl_id
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cpmk"
    post:
      operationId: createCpmk
      tags: [cpmk]
      summary: Tambah CPMK
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CpmkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cpmk"
  /cpmk/{id}:
    patch:
      operationId: updateCpmk
      tags: [cpmk]
      summary: Update CPMK
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CpmkUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cpmk"
    delete:
      operationId: deleteCpmk
      tags: [cpmk]
      summary: Hapus CPMK
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SUB-CPMK ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MataKuliah:
      type: object
      required: [id, kode, nama, sks, semester, tipe]
      properties:
        id:             { type: integer }
        kode:           { type: string }
        nama:           { type: string }
        sks:            { type: integer }
        semester:       { type: integer }
        tipe:           { type: string }
        kelompok:       { type: ["string", "null"] }
        dosenPengampu:  { type: ["string", "null"] }
        isAktif:        { type: boolean }
        mbkm:           { type: boolean }
        statusRps:      { type: ["string", "null"] }
        createdAt:      { type: string }
        updatedAt:      { type: string }
    MataKuliahInput:
      type: object
      required: [kode, nama, sks, semester, tipe]
      properties:
        kode:           { type: string }
        nama:           { type: string }
        sks:            { type: integer, minimum: 1, maximum: 6 }
        semester:       { type: integer, minimum: 1, maximum: 8 }
        tipe:           { type: string }
        kelompok:       { type: string }
        dosenPengampu:  { type: string }
        isAktif:        { type: boolean }
        mbkm:           { type: boolean }
    MataKuliahUpdate:
      type: object
      properties:
        kode:           { type: string }
        nama:           { type: string }
        sks:            { type: integer, minimum: 1, maximum: 6 }
        semester:       { type: integer, minimum: 1, maximum: 8 }
        tipe:           { type: string }
        kelompok:       { type: string }
        dosenPengampu:  { type: string }
        isAktif:        { type: boolean }
        mbkm:           { type: boolean }

    # ── CPL Prodi ──────────────────────────────────────────────────────────────
    CplProdi:
      type: object
      required: [id, kode, deskripsi]
      properties:
        id:          { type: integer }
        kode:        { type: string }
        deskripsi:   { type: string }
        jumlahMk:    { type: integer }
        jumlahCpmk:  { type: integer }
        createdAt:   { type: string }
        updatedAt:   { type: string }
    CplProdiInput:
      type: object
      required: [kode, deskripsi]
      properties:
        kode:       { type: string }
        deskripsi:  { type: string }
    CplProdiUpdate:
      type: object
      properties:
        kode:       { type: string }
        deskripsi:  { type: string }

    # ── CPMK ──────────────────────────────────────────────────────────────────
    Cpmk:
      type: object
      required: [id, kode, cplProdiId, deskripsi]
      properties:
        id:           { type: integer }
        kode:         { type: string }
        cplProdiId:   { type: integer }
        cplKode:      { type: ["string", "null"] }
        deskripsi:    { type: string }
        createdAt:    { type: string }
        updatedAt:    { type: string }
    CpmkInput:
      type: object
      required: [kode, cplProdiId, deskripsi]
      properties:
        kode:        { type: string }
        cplProdiId:  { type: integer }
        deskripsi:   { type: string }
    CpmkUpdate:
      type: object
      properties:
        kode:        { type: string }
        cplProdiId:  { type: integer }
        deskripsi:   { type: string }

    # ── Sub-CPMK ──────────────────────────────────────────────────────────────
...